requirements.txt
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
frozenlist==1.3.3
idna==3.4
loguru==0.6.0
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
pycparser==2.21
pyspnego==0.8.0
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
six==1.16.0
smbprotocol==1.10.1
soupsieve==2.4
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
yarl==1.8.2
